Struct c_ares_resolver::FutureResolver

pub struct FutureResolver { /* fields omitted */ }

An asynchronous DNS resolver, which returns results as futures::Futures.

Methods

impl FutureResolver

fn new() -> Result<FutureResolver, Error>

Create a new FutureResolver, using default Options.

fn with_options(options: Options) -> Result<FutureResolver, Error>

Create a new FutureResolver, with the given Options.

fn set_servers(&mut self, servers: &[&str]) -> Result<&mut Self, Error>

Set the list of servers to contact, instead of the servers specified in resolv.conf or the local named.

String format is host[:port]. IPv6 addresses with ports require square brackets eg [2001:4860:4860::8888]:53.

fn set_local_ipv4(&mut self, ipv4: &Ipv4Addr) -> &mut Self

Set the local IPv4 address from which to make queries.

fn set_local_ipv6(&mut self, ipv6: &Ipv6Addr) -> &mut Self

Set the local IPv6 address from which to make queries.

fn set_local_device(&mut self, device: &str) -> &mut Self

Set the local device from which to make queries.

fn query_a(&self, name: &str) -> CAresFuture<AResults>

Look up the A records associated with name.

fn search_a(&self, name: &str) -> CAresFuture<AResults>

Search for the A records associated with name.

fn query_aaaa(&self, name: &str) -> CAresFuture<AAAAResults>

Look up the AAAA records associated with name.

fn search_aaaa(&self, name: &str) -> CAresFuture<AAAAResults>

Search for the AAAA records associated with name.

fn query_cname(&self, name: &str) -> CAresFuture<CNameResults>

Look up the CNAME records associated with name.

fn search_cname(&self, name: &str) -> CAresFuture<CNameResults>

Search for the CNAME records associated with name.

fn query_mx(&self, name: &str) -> CAresFuture<MXResults>

Look up the MX records associated with name.

fn search_mx(&self, name: &str) -> CAresFuture<MXResults>

Search for the MX records associated with name.

fn query_naptr(&self, name: &str) -> CAresFuture<NAPTRResults>

Look up the NAPTR records associated with name.

fn search_naptr(&self, name: &str) -> CAresFuture<NAPTRResults>

Search for the NAPTR records associated with name.

fn query_ns(&self, name: &str) -> CAresFuture<NSResults>

Look up the NS records associated with name.

fn search_ns(&self, name: &str) -> CAresFuture<NSResults>

Search for the NS records associated with name.

fn query_ptr(&self, name: &str) -> CAresFuture<PTRResults>

Look up the PTR records associated with name.

fn search_ptr(&self, name: &str) -> CAresFuture<PTRResults>

Search for the PTR records associated with name.

fn query_soa(&self, name: &str) -> CAresFuture<SOAResult>

Look up the SOA records associated with name.

fn search_soa(&self, name: &str) -> CAresFuture<SOAResult>

Search for the SOA records associated with name.

fn query_srv(&self, name: &str) -> CAresFuture<SRVResults>

Look up the SRV records associated with name.

fn search_srv(&self, name: &str) -> CAresFuture<SRVResults>

Search for the SRV records associated with name.

fn query_txt(&self, name: &str) -> CAresFuture<TXTResults>

Look up the TXT records associated with name.

fn search_txt(&self, name: &str) -> CAresFuture<TXTResults>

Search for the TXT records associated with name.

fn cancel(&mut self)

Cancel all requests made on this FutureResolver.